+++ src/components/dashboard/Send.js
@@ -1,7 +1,5 @@
 import React, { useEffect, useState } from 'react'
 import isEmail from 'validator/lib/isEmail'
-import { withStyles } from '../../lib/styles'
-import normalize from '../../lib/utils/normalizeText'
 import { BackButton, useScreenState } from '../appNavigation/stackNavigation'
 import userStorage from '../../lib/gundb/UserStorage'
 import logger from '../../lib/logger/pino-logger'
@@ -137,16 +135,8 @@ const Send = props => {
   )
 }
 
-const getStylesFromProps = ({ theme }) => ({
-  cancelButton: {
-    color: theme.colors.gray80Percent,
-    fontSize: normalize(14),
-    fontWeight: '500',
-  },
-})
-
 Send.navigationOptions = {
   title: SEND_TITLE,
 }
 
-export default withStyles(getStylesFromProps)(Send)
+export default Send
